move the controller code to pkg/
